+++ packages/remax-cli/src/__tests__/integration/fixtures/runtime-plugin/expected/remax-vendors.js
@@ -67,6 +67,8 @@ __webpack_require__.r(__webpack_exports__);
 
 var ReactReconcilerInst = react_reconciler__WEBPACK_IMPORTED_MODULE_0___default()(_hostConfig__WEBPACK_IMPORTED_MODULE_1__["default"]);
 
+if (false) {}
+
 function getPublicRootInstance(container) {
   var containerFiber = container.current;
 
@@ -638,9 +640,7 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "propAlias", function() { return propAlias; });
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "default", function() { return propsAlias; });
 /* harmony import */ var _utils_plainStyle__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(11);
-/* harmony import */ var _RuntimeOptions__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(13);
-function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; } return _typeof(obj); }
-
+/* harmony import */ var _createHostComponent__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(13);
 
 
 function getAlias(prop, type) {
@@ -658,7 +658,7 @@ function getAlias(prop, type) {
 }
 
 function getValue(prop, value) {
-  if (prop.toLowerCase().endsWith('style') && _typeof(value) === 'object') {
+  if (prop.toLowerCase().endsWith('style') && Object.prototype.toString.call(value) === '[object Object]') {
     return Object(_utils_plainStyle__WEBPACK_IMPORTED_MODULE_0__["default"])(value);
   }
 
@@ -1094,7 +1094,7 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "DEPRECATED_CATCH_TYPE", function() { return DEPRECATED_CATCH_TYPE; });
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "SYNTHETIC_TYPES", function() { return SYNTHETIC_TYPES; });
 var DEPRECATED_CATCH_TYPE = 'catchClick';
-var SYNTHETIC_TYPES = ['onClick', 'onTap', 'onLongClick', 'onLongTap'];
+var SYNTHETIC_TYPES = ['onClick', 'onTap', 'onLongClick', 'onLongTap', 'onTouchMove', 'onTouchStart', 'onTouchEnd'];
 
 /***/ }),
 /* 18 */
@@ -1121,6 +1121,7 @@ var __assign = undefined && undefined.__assign || function () {
 
 var STYLE = 'style';
 var CHILDREN = 'children';
+var CLASS_NAME = 'className';
 function diffProperties(lastRawProps, nextRawProps) {
   var updatePayload = null;
   var lastProps = lastRawProps;
@@ -1149,7 +1150,7 @@ function diffProperties(lastRawProps, nextRawProps) {
     } else {
       // For all other deleted properties we add it to the queue. We use
       // the whitelist in the commit phase instead.
-      (updatePayload = updatePayload || []).push(propKey, null);
+      (updatePayload = updatePayload || []).push(propKey, propKey === CLASS_NAME ? '' : null);
     }
   }
 
@@ -2309,6 +2310,7 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);
 
 var PageInstanceContext = react__WEBPACK_IMPORTED_MODULE_0__["createContext"](null);
+PageInstanceContext.displayName = 'PageContext';
 /* harmony default export */ __webpack_exports__["default"] = (PageInstanceContext);
 
 /***/ }),
@@ -2380,6 +2382,23 @@ var __spread = undefined && undefined.__spread || function () {
   return ar;
 };
 
+var __values = undefined && undefined.__values || function (o) {
+  var s = typeof Symbol === "function" && Symbol.iterator,
+      m = s && o[s],
+      i = 0;
+  if (m) return m.call(o);
+  if (o && typeof o.length === "number") return {
+    next: function next() {
+      if (o && i >= o.length) o = void 0;
+      return {
+        value: o && o[i++],
+        done: !o
+      };
+    }
+  };
+  throw new TypeError(s ? "Object is not iterable." : "Symbol.iterator is not defined.");
+};
+
 
 
 
@@ -2390,6 +2409,7 @@ var Container =
 function () {
   function Container(context) {
     this.updateQueue = [];
+    this.deletedPaths = new Set();
     this.rendered = false;
     this.context = context;
     this.root = new _VNode__WEBPACK_IMPORTED_MODULE_0__["default"]({
@@ -2404,6 +2424,10 @@ function () {
     var _this = this;
 
     if (immediately) {
+      if (update.deleteCount === 1) {
+        this.deletedPaths.add(update.path);
+      }
+
       this.updateQueue.push(update);
       this.applyUpdate();
     } else {
@@ -2411,6 +2435,10 @@ function () {
         Promise.resolve().then(function () {
           return _this.applyUpdate();
         });
+      }
+
+      if (update.deleteCount === 1) {
+        this.deletedPaths.add(update.path);
       }
 
       this.updateQueue.push(update);
@@ -2462,14 +2490,36 @@ function () {
         });
       });
       this.updateQueue = [];
+      this.deletedPaths.clear();
       return;
     }
 
     var updatePayload = this.updateQueue.reduce(function (acc, update) {
-      var _a, _b;
+      var e_1, _a, _b, _c;
+
+      try {
+        // 如果父元素已经删除了，跳过所有对其子元素的操作
+        for (var _d = __values(_this.deletedPaths), _e = _d.next(); !_e.done; _e = _d.next()) {
+          var deletedPath = _e.value;
+
+          if (new RegExp("^" + deletedPath + ".nodes").test(update.path)) {
+            return acc;
+          }
+        }
+      } catch (e_1_1) {
+        e_1 = {
+          error: e_1_1
+        };
+      } finally {
+        try {
+          if (_e && !_e.done && (_a = _d.return)) _a.call(_d);
+        } finally {
+          if (e_1) throw e_1.error;
+        }
+      }
 
       if (update.type === 'splice') {
-        var item = __assign(__assign({}, acc), (_a = {}, _a[update.path + '.nodes.' + update.id] = update.items[0] || null, _a));
+        var item = __assign(__assign({}, acc), (_b = {}, _b[update.path + '.nodes.' + update.id] = update.items[0] || null, _b));
 
         if (update.children) {
           item[update.path + '.children'] = (update.children || []).map(function (c) {
@@ -2480,7 +2530,7 @@ function () {
         return item;
       }
 
-      return __assign(__assign({}, acc), (_b = {}, _b[update.path + '.' + update.name] = update.value, _b));
+      return __assign(__assign({}, acc), (_c = {}, _c[update.path + '.' + update.name] = update.value, _c));
     }, {});
     this.context.setData(updatePayload, function () {
       _nativeEffect__WEBPACK_IMPORTED_MODULE_2__["default"].run();
@@ -2491,6 +2541,7 @@ function () {
       }
     });
     this.updateQueue = [];
+    this.deletedPaths.clear();
   };
 
   Container.prototype.clearUpdate = function () {