+++ packages/remax-runtime/typings/index.d.ts
@@ -4,3 +4,4 @@ declare const getCurrentPages: any;
 declare const getApp: any;
 
 declare module '@remax/runtime-plugin';
+declare module 'react-devtools-core';